+++ src/modules/variable.js
@@ -1,0 +1,6 @@
+
+const taskContainer = document.querySelector(".task-container");
+const taskInput = document.querySelector(".task-input");
+const clearComplete = document.querySelector(".clear-completed");
+
+export {taskContainer, taskInput, clearComplete}
